Put call quality in Prometheus and behind a dashboard
RTT, packet loss and jitter reached the browser's stats report and then stopped
at a log line and an in-memory snapshot, so the only way to see them was to be
watching when a call happened. ObserveMedia records them as histograms, which
makes the numbers queryable after the fact and gives the alert rules something
to evaluate.

Bitrates were typed as int while the browser sends fractions of a kbps, and
strict decoding rejected the whole report on the first fractional value. The
stats-report handler also swallowed decode errors, so a rejected report looked
identical to a report that never arrived and network-quality counts sat at
zero. Both are float64 now and a malformed report returns instead of
half-populating a peer entry.

deploy/monitoring carries the Prometheus and Grafana images with a dashboard
covering call setup, repair outcomes, network quality, API latency and server
resources, plus five alert rules and 14 days of retention.

+++ internal/signaling/message.go
@@ -135,8 +135,8 @@ type StatsReportPeer struct {
 	NetworkPressure     string  `json:"networkPressure"`
 	RoundTripTimeMs     float64 `json:"roundTripTimeMs"`
 	PacketLossPercent   float64 `json:"packetLossPercent"`
-	OutboundBitrateKbps int     `json:"outboundBitrateKbps"`
-	InboundBitrateKbps  int     `json:"inboundBitrateKbps"`
+	OutboundBitrateKbps float64 `json:"outboundBitrateKbps"`
+	InboundBitrateKbps  float64 `json:"inboundBitrateKbps"`
 	CandidateType       string  `json:"candidateType"`
 	JitterMs            float64 `json:"jitterMs"`
 	EncodingLevel       int     `json:"encodingLevel"`
